moved fake trivia data into another file
