+++ client/src/fakeData.js
@@ -1,0 +1,81 @@
+const trivias = [
+  {
+    title: "Marvek ",
+    image:
+      "https://images.unsplash.com/photo-1593642634443-44adaa06623a?ixid=MnwxMjA3fDF8MHxlZGl0b3JpYWwtZmVlZHw2fHx8ZW58MHx8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=800&q=60",
+    likes: 5,
+    tags: ["funny", "difficult", "dsfs"],
+  },
+  {
+    title: "Marvel ",
+    image:
+      "https://images.unsplash.com/photo-1635705830341-052af8af5426?ixid=MnwxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHwzfHx8ZW58MHx8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=500&q=60",
+    likes: 235,
+    tags: ["hero", "difficult", "200IQ"],
+  },
+  {
+    title: "Marsdf ",
+    image:
+      "https://images.unsplash.com/photo-1635701680062-b53d105b1088?ixid=MnwxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHwxN3x8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=500&q=60",
+    likes: 5,
+    tags: ["funny", "dier", "200IQ"],
+  },
+  {
+    title: "Marvdsfek ",
+    image:
+      "https://images.unsplash.com/photo-1635703256356-616cfaf96398?ixid=MnwxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHwyMHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=500&q=60",
+    likes: 5,
+    tags: ["funny", "difficult", "Brigs"],
+  },
+  {
+    title: "Marvdsfek ",
+    image:
+      "https://images.unsplash.com/photo-1635703256356-616cfaf96398?ixid=MnwxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHwyMHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=500&q=60",
+    likes: 5,
+    tags: ["funny", "difficult", "Brigs"],
+  },
+  {
+    title: "Marvdsfek ",
+    image:
+      "https://images.unsplash.com/photo-1635703256356-616cfaf96398?ixid=MnwxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHwyMHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=500&q=60",
+    likes: 5,
+    tags: ["funny", "difficult", "Brigs"],
+  },
+  {
+    title: "Marvdsfek ",
+    image:
+      "https://images.unsplash.com/photo-1635703256356-616cfaf96398?ixid=MnwxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHwyMHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=500&q=60",
+    likes: 5,
+    tags: ["funny", "difficult", "Brigs"],
+  },
+  {
+    title: "Marvdsfek ",
+    image:
+      "https://images.unsplash.com/photo-1635703256356-616cfaf96398?ixid=MnwxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHwyMHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=500&q=60",
+    likes: 5,
+    tags: ["funny", "difficult", "Brigs"],
+  },
+  {
+    title: "Marvdsfek ",
+    image:
+      "https://images.unsplash.com/photo-1635703256356-616cfaf96398?ixid=MnwxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHwyMHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=500&q=60",
+    likes: 5,
+    tags: ["funny", "difficult", "Brigs"],
+  },
+  {
+    title: "Marvdsfek ",
+    image:
+      "https://images.unsplash.com/photo-1635703256356-616cfaf96398?ixid=MnwxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHwyMHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=500&q=60",
+    likes: 5,
+    tags: ["funny", "difficult", "Brigs"],
+  },
+  {
+    title: "Marvdsfek ",
+    image:
+      "https://images.unsplash.com/photo-1635703256356-616cfaf96398?ixid=MnwxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHwyMHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=500&q=60",
+    likes: 5,
+    tags: ["funny", "difficult", "Brigs"],
+  },
+];
+
+export default trivias;
